# wallbox/wallbox_mobile/API-DOC/swagger2.yml
openapi: 3.0.1
info:
  title: Wallbox Project API
  description: API for Wallbox mobile applications (Odoo v18 backend)
  version: "1.1.0"
  contact:
    name: API Support
    email: munafkhan732@gmail.com
servers:
  - url: https://qariqa.qariqa.com/v1
    description: Production server
  - url: https://qariqa-staging.qariqa.com/v1
    description: Staging server
...
paths:
...
  /wallboxes/{wallbox_id}/invitations:
    parameters:
      - $ref: '#/components/parameters/WallboxId'
    get:
      tags: [Sharing & Invitations]
      summary: List invitations
      description: Lists all users who have been invited to use a specific wallbox.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of invitations.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags: [Sharing & Invitations]
      summary: Create invitation
      description: Invites a user (by email) to use a specific wallbox.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvitationCreateRequest'
      responses:
        '201':
          description: Invitation sent successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /wallboxes/{wallbox_id}/invitations/{invite_id}:
    parameters:
      - $ref: '#/components/parameters/WallboxId'
      - $ref: '#/components/parameters/InviteId'
    put:
      tags: [Sharing & Invitations]
      summary: Update invitation
      description: Updates the details of an invited user.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvitationUpdateRequest'
      responses:
        '200':
          description: Invitation updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      tags: [Sharing & Invitations]
      summary: Delete invitation
      description: Revokes a user's access to a specific wallbox.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Invitation deleted successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

  # Send Invitations (Owner to User)
...
components:
  schemas:
...
    Invitation:
      type: object
      properties:
        invite_id:
          type: integer
        email:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [pending, active]
        access_type:
          type: string
          example: monthly
    InvitationCreateRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        phone:
          type: string
      required: [email]
    InvitationUpdateRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
    InvitationResponse:
      type: object
      properties:
        status:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/Invitation'
    InvitationListResponse:
      type: object
      properties:
        status:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Invitation'
        count:
          type: integer

    # Send Invitation Schemas
...
    SuccessResponse:
      type: object
      properties:
        status:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation was successful"
    ErrorResponse:
      type: object
      properties:
        status:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              description: Additional error details
              nullable: true
      required: [status, error]
  parameters:
...
    WallboxId:
      name: wallbox_id
      in: path
      required: true
      schema:
        type: integer
      description: Wallbox (Charging Station) ID
    InviteId:
      name: invite_id
      in: path
      required: true
      schema:
        type: integer
      description: Invitation (Partner) ID
  responses:
    BadRequest:
      description: Bad request (invalid input)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: false
            error:
              code: "VALIDATION_ERROR"
              message: "Required fields are missing"
    Unauthorized:
      description: Unauthorized (invalid or missing token)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: false
            error:
              code: "TOKEN_INVALID"
              message: "Invalid or expired authentication token"
    Forbidden:
      description: Forbidden (insufficient permissions)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: false
            error:
              code: "ACCESS_DENIED"
              message: "You don't have permission to perform this action"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: false
            error:
              code: "RESOURCE_NOT_FOUND"
              message: "The requested resource was not found"
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: false
            error:
              code: "INTERNAL_ERROR"
              message: "An unexpected error occurred"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT